pyproject.toml
[tool.poetry]
name = "cognite-extractor-utils-mqtt"
version = "0.1.0"
description = "MQTT extension for the Cognite extractor-utils framework"
authors = ["Einar Omang <einar.omang@cognite.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/cognitedata/python-extractor-utils-mqtt"

packages = [
    { include="cognite/extractorutils/mqtt", from="." },
]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["arrow", "dacite", "more_itertools", "requests"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
cognite-extractor-utils = "^2.0.0"
more-itertools = "^8.12.0"
typing-extensions = "^4.0.1"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
mypy = "*"
flake8 = "*"
pytest = "^6.2.5"
pytest-cov = "^2.8.1"
sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"
pre-commit = "^2.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
